% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_enrichResult.R
\name{as_enrichResult}
\alias{as_enrichResult}
\title{Convert enrichr_all to enrichResults}
\usage{
as_enrichResult(e1, db = 1, vs = 1, genes)
}
\arguments{
\item{e1}{results from enrichr_all}

\item{db}{List index or name, default 1}

\item{vs}{Contrast number or name, default 1}

\item{genes}{vector of significant genes tested}
}
\value{
enrichResults object
}
\description{
https://github.com/jmw86069/multienrichjam/blob/629b347911ae3306d8a6ae1007668ec86a9a292f/R/jamenrich-base.r
}
\examples{
\dontrun{
library(enrichR)
setEnrichrSite("Enrichr")
celldbs <- c("KEGG_2021_Human", "WikiPathways_2019_Human")
e1 <- enrichr_all(res, celldbs)
sig <- sort(unique(res[[2]]$gene_name[res[[2]]$padj < 0.05]))
e2 <- as_enrichResult(e1, 1,2, genes = sig)
library(ClusterProfiler)
barplot(e2, show=10)
dotplot(e2)
heatplot(e2, show=10)
library(enrichplot)
e2 <- pairwise_termsim(e2)
emapplot(e2, show=10)
emapplot(e2, show=10, cex_line=3) + ggraph::scale_edge_width(range = c(1,10))
}
}
\author{
Chris Stubben
}
