% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ipa.R
\name{barplot_ipa}
\alias{barplot_ipa}
\title{Barplot of top pathways, regulators, disease or toxicity functions.}
\usage{
barplot_ipa(
  ipa,
  top = 15,
  pvalue = 0.05,
  wrap = 30,
  barwidth = 0.7,
  name_category = FALSE,
  stacked = FALSE,
  results
)
}
\arguments{
\item{ipa}{a table from read_IPA}

\item{top}{Number of top pathways, regulators or functions to display, default 15}

\item{pvalue}{P-value cutoff, default 0.05}

\item{wrap}{wrap pathway strings, default 40 characters}

\item{barwidth}{geom_bar width, default 0.7}

\item{name_category}{add Category to x-axis name, default FALSE}

\item{stacked}{plot stacked barchart, default FALSE}

\item{results}{Analysis Ready Molecules from read_ipa, required for stacked pathway barplot}
}
\value{
a ggplot
}
\description{
Bars are colored by z-score and height is p-value.  If stacked=TRUE, plot the
percent of up and down-regulated genes or activated and inhibited regulators by Category
}
\examples{
\dontrun{
ipa1 <- read_ipa("IPA_export_all.txt")
barplot_ipa(ipa1[[2]])
barplot_ipa(ipa1[[2]], stacked=TRUE, results=ipa1[[10]])
barplot_ipa(ipa1[[3]], name_cat=TRUE, wrap=30)
}
}
\author{
Chris Stubben
}
