% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_to_SAF.R
\name{biomart_to_SAF}
\alias{biomart_to_SAF}
\title{Rename biomart table to SAF format for featureCounts}
\usage{
biomart_to_SAF(x, gene_name = FALSE)
}
\arguments{
\item{x}{biomart table}

\item{gene_name}{use gene_name as GeneID, default Ensembl ID}
}
\value{
tibble
}
\description{
Rename biomart table to SAF format for featureCounts
}
\examples{
\dontrun{
}
}
\author{
Chris Stubben
}
