% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_ipa.R
\name{cluster_ipa}
\alias{cluster_ipa}
\title{Cluster genes in top IPA pathways}
\usage{
cluster_ipa(ipa, top = 10, cut = 50, ...)
}
\arguments{
\item{ipa}{a table from read_IPA}

\item{top}{Number of top pathways, regulators or functions to display, default 7}

\item{cut}{Cut long pathway names, default 50 characters}

\item{\dots}{additional options passed to scale_x_continuous}
}
\value{
a ggplot
}
\description{
Cluster genes in top IPA pathways
}
\examples{
\dontrun{
ipa1 <- read_ipa("IPA_export_all.txt")

}
}
\author{
Chris Stubben
}
