% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr_barplot.R
\name{enrichr_barplot}
\alias{enrichr_barplot}
\title{Plot enrichr_all output}
\usage{
enrichr_barplot(
  e1,
  vs = 1,
  width = 0.7,
  top = 10,
  wrap = 40,
  xlab_bar = "pathway",
  ...
)
}
\arguments{
\item{e1}{results from enrichr_all}

\item{vs}{Contrast number or name, default 1}

\item{width}{bar width}

\item{top}{Number of top sets to plot}

\item{wrap}{Wrap long string names, default 40}

\item{xlab_bar}{xlab name}

\item{\dots}{additional options passed to pheatmap}

\item{db}{List index or name, default 1}

\item{plot}{plot type, cluster, tile or default bar}

\item{top_n}{plot top sets in each contrast}

\item{cluster_rows}{cluster rows}

\item{exclude_row}{exclude rows}

\item{exclude_col}{exclude columns}

\item{set_first}{set as first}
}
\value{
plots
}
\description{
Plot enrichr_all output
}
\examples{
\dontrun{
library(enrichR)
setEnrichrSite("Enrichr")
celldbs <- c("KEGG_2021_Human", "WikiPathways_2019_Human")
e1 <- enrichr_all(res, celldbs)
plot_enrichr(e1, 1,2)
plot_enrichr(e1, 1,2, plot="cluster", fontsize_col=7)
}
}
\author{
Chris Stubben
}
