% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr_clusters.R
\name{enrichr_clusters}
\alias{enrichr_clusters}
\title{Loop through FindAllMarkers and run EnrichR}
\usage{
enrichr_clusters(clust, celldbs, padj = 0.1)
}
\arguments{
\item{clust}{results from FindAllMarkers with cluster and gene columns}

\item{celldbs}{a vector of library names from listEnrichrDbs}

\item{padj}{Adjusted P value cutoff for Enrichr results}
}
\value{
a list of tibbles
}
\description{
Loop through FindAllMarkers and run EnrichR
}
\examples{
\dontrun{
library(enrichR)
setEnrichrSite("Enrichr")
celldbs <- c("KEGG_2021_Human", "WikiPathways_2019_Human")
res <- FindAllMarkers(object = srt, only.pos = TRUE, logfc.threshold = 0.5)
e1 <- enrichr_clusters(res, celldbs)
}
}
\author{
Chris Stubben
}
