% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr_contrasts.R
\name{enrichr_contrasts}
\alias{enrichr_contrasts}
\alias{enrichr_gene}
\title{Loop through FindMarkers and run EnrichR}
\usage{
enrichr_contrasts(res, celldbs, padj = 0.1)

enrichr_gene(res, celldbs, padj = 0.1)
}
\arguments{
\item{res}{two or more results from FindMarkers, combined into a single table with contrast column}

\item{celldbs}{a vector of library names from listEnrichrDbs}

\item{padj}{Adjusted P value cutoff for Enrichr results}
}
\value{
a list of tibbles
}
\description{
Loop through FindMarkers and run EnrichR
}
\section{Functions}{
\itemize{
\item \code{enrichr_gene()}: Run Enrichr on single FindMarkers result

}}
\examples{
\dontrun{
library(enrichR)
setEnrichrSite("Enrichr")
celldbs <- c("KEGG_2021_Human", "WikiPathways_2019_Human")
e1 <- enrichr_contrasts(res, celldbs)
}
}
\author{
Chris Stubben
}
