% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_features.R
\name{eset_features}
\alias{eset_features}
\title{List features in Expression sets}
\usage{
eset_features(eset)
}
\arguments{
\item{eset}{expression set from \code{GEOquery}}
}
\value{
A tibble
}
\description{
List features in Expression sets
}
\author{
Chris Stubben
}
