% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_samples.R
\name{eset_samples}
\alias{eset_samples}
\title{List samples in Expression sets}
\usage{
eset_samples(eset, unique = FALSE)
}
\arguments{
\item{eset}{expression set from \code{GEOquery}}

\item{unique}{show columns with same repeated values}
}
\value{
A tibble
}
\description{
List samples in Expression sets
}
\author{
Chris Stubben
}
