% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_enrich_contrast.R
\name{multi_enrich_contrast}
\alias{multi_enrich_contrast}
\title{Plot one pathway in multiple contrasts}
\usage{
multi_enrich_contrast(
  x,
  fc,
  db,
  n = 4,
  ncol = 2,
  scaleplots = 0.95,
  sameyaxis = TRUE
)
}
\arguments{
\item{x}{a table from fgsea_all (top n rows are displayed)}

\item{fc}{fold change vector}

\item{db}{gene set database}

\item{n}{number of rows in x to display, default 4}

\item{ncol}{number of columns in plot_gird, default 2}

\item{scaleplots}{scale option for plot_grid, default 0.95}

\item{sameyaxis}{use the same y-axis range for all plots, default TRUE}
}
\value{
a plot_grid
}
\description{
Plot one pathway in multiple contrasts
}
\examples{
\dontrun{
k1 <- fgsea_all(res, msig_pathways$KEGG, FDR=1)
fc <- write_gsea_rnk(res, write=FALSE)
z <- bind_rows(k1, .id = "contrast") \%>\% filter(pathway=="Apoptosis")
multi_enrich_contrast(z, fc, msig_pathways$KEGG)
}
}
\author{
Chris Stubben
}
