% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_enrich_plot.R
\name{multi_enrich_plot}
\alias{multi_enrich_plot}
\title{Plot multiple enrichment plots}
\usage{
multi_enrich_plot(
  x,
  fc,
  db,
  n = 4,
  contrast,
  ncol = 2,
  scaleplots = 0.95,
  sameyaxis = TRUE
)
}
\arguments{
\item{x}{a table from fgsea_all (top n rows are displayed)}

\item{fc}{fold change vector}

\item{db}{gene set database}

\item{n}{number of rows in x to display, default 4}

\item{contrast}{add contrast name to title}

\item{ncol}{number of columns in plot_gird, default 2}

\item{scaleplots}{scale option for plot_grid, default 0.95}

\item{sameyaxis}{use the same y-axis range for all plots, default TRUE}
}
\value{
a plot_grid
}
\description{
Plot multiple enrichment plots
}
\examples{
\dontrun{
multi_enrich_plot(h1[[1]], fc[[1]], msig_hallmark, n=6)
}
}
\author{
Chris Stubben
}
