% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_chr.R
\name{pheatmap_chr}
\alias{pheatmap_chr}
\title{Plot heatmap for a single chromosome}
\usage{
pheatmap_chr(rld, biomart, gene_names, ...)
}
\arguments{
\item{rld}{rlog or other counts in DESeqTransform object}

\item{biomart}{filter annotations from \code{read_biomart}}

\item{gene_names}{a vector of gene names to label}

\item{\dots}{other options passed to \code{pheatmap}}
}
\value{
A pheatmap
}
\description{
Plot pheatmap for genes on a single chromosome and labels every 10, 20, 30 Mb
and optionally add a vector of gene names
}
\examples{
\dontrun{
 pheatmap_chr(rld,  subset(mmu, chromosome==11),  gene_names = "Ewsr1" )
}
}
\author{
Chris Stubben
}
