% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fisher.R
\name{plot_fisher}
\alias{plot_fisher}
\title{Plot Fisher all output}
\usage{
plot_fisher(
  x,
  trim = 70,
  sets,
  p.value = 0.05,
  percent = FALSE,
  palette = "Reds",
  cluster_row = FALSE,
  ...
)
}
\arguments{
\item{x}{list from \code{link{fgsea_all}}}

\item{trim}{trim long names, default more than 70 characters}

\item{sets}{display contrasts sharing n or more sets for n > 1.  If n = 1,
then only plot unique sets.  If missing, then plots all sets, default.}

\item{p.value}{pvalue p-value cutoff}

\item{percent}{plot percent in set, default is -log10(pvalue)}

\item{palette}{color palette, default Reds}

\item{cluster_row}{Cluster dendrogram rows, default is an alphabetical list}

\item{\dots}{other options passed to \code{pheatmap}}
}
\description{
Plot heatmap with p-values or percent by contrast
}
\examples{
\dontrun{
  library(hciRdata)
  x <- fisher_all(fc, msig_pathways$KEGG)
  plot_fisher(x)
}
}
\author{
Chris Stubben
}
