% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kegg.R
\name{plot_kegg}
\alias{plot_kegg}
\title{Plot KEGG pathway maps using pathview}
\usage{
plot_kegg(x, pathway_name)
}
\arguments{
\item{x}{a results table with gene_name and log2FoldChange}

\item{pathway_name}{match pathway name}
}
\description{
Plot KEGG pathways
}
\examples{
\dontrun{
  data(kegg_hsa)
  data(egSymb)   # in gage
  plot_kegg(res, "Cytokine")
 }
}
\author{
Chris Stubben
}
