% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot smoothed methylation}
\usage{
plot_methyl(
  bsseq,
  dmr,
  smooth = smoothPs,
  subset,
  group = "trt",
  type = c("a", "confint"),
  extend = 10000,
  ...
)
}
\arguments{
\item{bsseq}{bsseq results}

\item{dmr}{annotated dmrFinder output with 1 row (with start, end, and chr columns)}

\item{smooth}{smoothed methylation values}

\item{subset}{subset samples to plot}

\item{group}{column name for grouping, default trt}

\item{type}{track type}

\item{extend}{bases to extend up or downstream}

\item{\dots}{additional options passed to plotTracks}
}
\description{
Plot KEGG pathways
}
\examples{
\dontrun{
  plot_methyl
 }
}
\author{
Chris Stubben
}
