% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.R
\name{plot_QC}
\alias{plot_QC}
\title{Plot QC Concentration vs QC RIN}
\usage{
plot_QC(samples, ...)
}
\arguments{
\item{samples}{A GNomEx sample table with \code{QC RIN} and \code{QC Conc. (ng/uL)}}

\item{\dots}{additional options passed to hc_chart}
}
\value{
A highchart
}
\description{
Plot QC Concentration vs QC RIN
}
\examples{
\dontrun{
plot_QC(samples, x="QC Conc. (ng/uL)", y="QC RIN")
}
}
\author{
Chris Stubben
}
