% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_UMIs.R
\name{plot_UMIs}
\alias{plot_UMIs}
\title{Plot UMIs found by qiagen_smallRNA_umi_extractor}
\usage{
plot_UMIs(x, stack = "percent", width = 15, ...)
}
\arguments{
\item{x}{file with sample id, UMIs found and no UMI tags}

\item{stack, }{plot percent or total reads (percent or normal)}

\item{width}{bar width, default 15}

\item{\dots}{additional options passed to \code{hc_size}}
}
\value{
A highchart
}
\description{
Plot UMIs found by qiagen_smallRNA_umi_extractor
}
\examples{
\dontrun{
 x <- read_tsv("UMIs.txt")
 plot_UMIs(x)
}
}
\author{
Chris Stubben
}
