% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vargenes.R
\name{plot_vargenes}
\alias{plot_vargenes}
\title{Bar plot of variance fractions for top genes}
\usage{
plot_vargenes(x, column, n = 20, barwidth = NULL, first = FALSE, ...)
}
\arguments{
\item{x}{result from fitExtractVarPartModel in variancePartition package}

\item{column}{column name}

\item{n}{number of top variable genes, default 20}

\item{barwidth}{barplot pointWidth}

\item{first}{put column first in key}

\item{\dots}{additional options passed to \code{hc_size}}
}
\value{
A highchart
}
\description{
Bar plot of variance fractions for top genes
}
\examples{
\dontrun{
 x <- read_tsv("biotypes.txt")
 plot_biotypes(x)
}
}
\author{
Chris Stubben
}
