% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{Read all Excel spreadsheets}
\usage{
read_excel_sheets(file)
}
\arguments{
\item{file}{Excel file name}
}
\value{
A list of tibbles
}
\description{
Read all Excel spreadsheets
}
\examples{
\dontrun{
   read_excel_files()
}
}
\author{
Chris Stubben
}
