% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seal.R
\name{read_seal}
\alias{read_seal}
\title{Read SEAL count files and merge into a single matrix}
\usage{
read_seal(path = ".", pattern, reshape = TRUE)
}
\arguments{
\item{path}{the path to SEAL output files, the default
corresponds to the working directory.}

\item{pattern}{regular expression for file name matching, default .txt}

\item{reshape}{reshape into wide format with samples in rows (count matrix)}
}
\value{
A tibble
}
\description{
Read SEAL count files and merge into a single matrix
}
\examples{
\dontrun{
   read_seal()
}
}
\author{
Chris Stubben
}
