% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sjout.R
\name{read_sjout}
\alias{read_sjout}
\title{Read STAR SJ.out  file}
\usage{
read_sjout(file)
}
\arguments{
\item{file}{file name}
}
\value{
tibble
}
\description{
Read STAR SJ.out  file
}
\examples{
\dontrun{
}
}
\author{
Chris Stubben
}
