% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_coverage_starts.R
\name{riboseq_coverage_starts}
\alias{riboseq_coverage_starts}
\title{Estimate coverage of riboseq read start around start and stop codon}
\usage{
riboseq_coverage_starts(cds_by_tx, aln, average = TRUE, name = "Ribo-seq")
}
\arguments{
\item{cds_by_tx}{A GRangesList from \code{cdsBy(txdb, by="tx")} with first
and last exon extended 50 bases using \code{resizeFeature} from
\code{systemPipeR}}

\item{aln}{GAlignments object from \code{readGAlignments}}

\item{average}{return entire coverage vector as RleList instead of average
coverage around start and stop codon}

\item{name}{sample name column value}
}
\value{
A tibble with start, mean start coverage (avg_start), number of
genes with at least one read (n_start), frame, end, avg_end, sample
name and number of transcripts
}
\description{
Finds coverage of the first base in the alignment between -50 and 50 bp
around the start codon and between -50 and 50 bp around stop codon.
}
\examples{
\dontrun{
 library(GenomicAlignments)
 aln1 <- readGAlignments(BamFile("uniq_x1.bam"))
 library(GenomicFeatures)
 txdb <- makeTxDbFromEnsembl(organism="Mus musculus", release=102)
 seqlevels(txdb) <- seqlevels(aln1)
 cds_by_tx <- cdsBy(txdb, by="tx", use.names=TRUE)
 cds_by_tx50 <- systemPipeR:::.resizeFeature(cds_by_tx, up=50, down=50)
 a1 <- subset(aln1, width(aln1) >= 27 & width(aln1) <= 30 )
 x1 <- riboseq_coverage_starts( cds_by_tx50, a1)
 x1
}
}
\author{
Chris Stubben
}
