% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_psite_offsets.R
\name{riboseq_psite_offsets}
\alias{riboseq_psite_offsets}
\title{Estimate riboseq coverage for different read sizes.}
\usage{
riboseq_psite_offsets(lengths, cds, aln)
}
\arguments{
\item{lengths}{a vector of alignment lengths}

\item{cds}{A GRangesList from \code{cdsBy(txdb, by="tx")} with first and last exon
extended 50 bases using \code{resizeFeature} from \code{systemPipeR}}

\item{aln}{GAlignments object from \code{readGAlignments}}
}
\value{
A tibble with read length and columns from code{riboseq_coverage_start}
}
\description{
Loops through an alignment file and runs code{riboseq_coverage_start}
}
\examples{
\dontrun{
 psite1 <- psite_offsets(26:31, cds_by_tx50, aln1)
 dplyr::select(psite1, 1,8) \%>\% unique()
 ggplot(psite1, aes(x=start, y=avg_start, color=length)) + geom_line() +
 xlab("Start of read relative to start codon") +
 ylab("Reads per transcript") + theme(legend.position="none") +
 facet_wrap( ~ length, ncol=1, scales="free", strip.position="right")
}
}
\author{
Chris Stubben
}
