% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlog_diff.R
\name{rlog_diff}
\alias{rlog_diff}
\title{Calculate fold changes using rlog differences}
\usage{
rlog_diff(rld, exp, ref, biomart)
}
\arguments{
\item{rld}{a DESeqTransform object with regularized log values}

\item{exp}{column numbers for treatment}

\item{ref}{column numbers for reference level, will be repeated if needed}

\item{biomart}{annotations from read_biomart}
}
\value{
a list of result tables with log2FoldChange
}
\description{
Calculate fold changes using rlog differences
}
\examples{
\dontrun{
  rlog_diff(rld, "G1", mouse96)
}
}
\author{
Chris Stubben
}
