% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_annotate.R
\name{slice_annotate}
\alias{slice_annotate}
\title{Add annotations to slice_with_height output}
\usage{
slice_annotate(gr1, biomart)
}
\arguments{
\item{gr1}{slice_to_height output}

\item{biomart}{biomart annotations}
}
\value{
A tibble
}
\description{
Finds nearest feature using distanceToNearest and adds gene and biotype.
The distance to region will be zero if peak is within the feature.
}
\examples{
\dontrun{
  rbga <- readGAlignments("14415X1.bam")
  slices <- slice_with_height(rgba, lower = 1000)
   slice_annotate(slices, human102)
}
}
\author{
Chris Stubben
}
