% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_to_tibble.R
\name{slice_to_tibble}
\alias{slice_to_tibble}
\title{Slice a coverage vector and convert to tibble}
\usage{
slice_to_tibble(cov, lower = 100)
}
\arguments{
\item{cov}{Coverage RleList}

\item{lower}{lower bound of slice}
}
\value{
A tibble with peak width and height
}
\description{
Converts the IRanges output from slice to a tibble with peak widths and
heights
}
\examples{
\dontrun{
  rbga <- readGAlignments("14415X1.bam")
  bamCov <- coverage(rbga)
  slice_to_tibble(bamCov, lower = 100)
}
}
\author{
Chris Stubben
}
