% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_with_height.R
\name{slice_with_height}
\alias{slice_with_height}
\title{Slice a stranded genome alignment and add peak height}
\usage{
slice_with_height(aln, lower = 100, min_width = 10)
}
\arguments{
\item{aln}{a GAlignments object}

\item{lower}{lower bound of slice}

\item{min_width}{remove peaks with 10 or fewer bases}
}
\value{
A GRanges
}
\description{
Converts slice output to a GRanges object and adds height using viewMaxs
}
\examples{
\dontrun{
  rbga <- readGAlignments("14415X1.bam")
  slice_with_height(rbga, lower = 100)
}
}
\author{
Chris Stubben
}
