% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tibble.R
\name{top_tibble}
\alias{top_tibble}
\alias{top_tibble_all}
\alias{write_top}
\title{Tibble of top genes from linear fit model with limma}
\usage{
top_tibble(fit, coef = 1, genes_columns)

top_tibble_all(fit, ...)

write_top(res, cutoff = 0.05)
}
\arguments{
\item{fit}{results from \code{lmFit}}

\item{coef}{contrast number}

\item{genes_columns}{column names in \code{fit$genes}, default is Gene.Symbol
and Gene.Title}

\item{\dots}{other options passed to \code{top_tibble}}

\item{res}{result for write_top}

\item{cutoff}{cutoff for write_top}
}
\value{
A list of tibbles
}
\description{
Tibble of top genes from linear fit model with limma
}
\section{Functions}{
\itemize{
\item \code{top_tibble_all()}: All top tibbles

\item \code{write_top()}: Write tables to file

}}
\examples{
\dontrun{
}
}
\author{
Chris Stubben
}
